backend/pyproject.toml
[project]
name = "discredit"
version = "0.1.0"
description = "Community intelligence mining system for Discord and Reddit"
authors = [
    {name = "ul0gic", email = "ul0gic@corelift.io"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "praw (>=7.8.1,<8.0.0)",
    "chromadb (>=1.3.4,<2.0.0)",
    "neo4j (>=6.0.3,<7.0.0)",
    "openai (>=2.7.1,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "ratelimit (>=2.2.1,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
